Add unlimited bandwidth helper and split parse tests
